app/graphql/router.py
from fastapi import APIRouter, HTTPException
from strawberry.fastapi import GraphQLRouter
from app.graphql.schema import schema

router = APIRouter()
graphql_app = GraphQLRouter(schema)
router.include_router(graphql_app, prefix="/graphql", tags=["GraphQL"])


