requirements.txt
fastapi
xmltodict
python-dotenv
supabase
openai
uvicorn
pydantic
requests
pytest
pytest-mock
dotenv
strawberry-graphql
black
flake8